refactor(html): tree construction
