+++ html/parser/tree_construction/rules/head.rs
@@ -3,10 +3,11 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 use html_elements::{interface::IsOneOfTagsInterface, tag_names};
+use infra::namespace::Namespace;
 
 use crate::{
-    state::InsertionMode,
-    tokenization::{HTMLTagToken, HTMLToken},
+    state::{FramesetOkFlag, InsertionMode, ScriptingFlag},
+    tokenization::{HTMLTagToken, HTMLToken, HTMLTokenizerState},
     tree_construction::{
         HTMLTreeConstruction, HTMLTreeConstructionControlFlow,
     },
